[refactor]:move all classes related to TiDB to ecnu.db.tidb
