+++ src/main/java/ecnu/db/Main.java
@@ -5,17 +5,17 @@ import com.alibaba.druid.util.JdbcConstants;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import ecnu.db.analyzer.online.AbstractAnalyzer;
-import ecnu.db.analyzer.online.TidbAnalyzer;
+import ecnu.db.tidb.TidbAnalyzer;
 import ecnu.db.analyzer.online.node.ExecutionNode;
 import ecnu.db.constraintchain.chain.ConstraintChain;
 import ecnu.db.constraintchain.filter.Parameter;
 import ecnu.db.dbconnector.AbstractDbConnector;
 import ecnu.db.dbconnector.DatabaseConnectorInterface;
 import ecnu.db.dbconnector.DumpFileConnector;
-import ecnu.db.dbconnector.TidbConnector;
+import ecnu.db.tidb.TidbConnector;
 import ecnu.db.schema.Schema;
-import ecnu.db.schema.generation.AbstractSchemaGenerator;
-import ecnu.db.schema.generation.TidbSchemaGenerator;
+import ecnu.db.schema.AbstractSchemaGenerator;
+import ecnu.db.tidb.TidbSchemaGenerator;
 import ecnu.db.utils.*;
 import ecnu.db.exception.TouchstoneToolChainException;
 import ecnu.db.exception.UnsupportedDatabaseSourceException;
